+++ test/unit/core/MakerPSMUnitTest.t.sol
@@ -14,11 +14,11 @@ import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 
 contract MakerPSMDummy is MakerPSM {
     IPSM psm;
-    IERC20 gem;
+    IERC20 dai;
 
-    constructor(IPSM _psm, IERC20 _gem) {
+    constructor(IPSM _psm, IERC20 _dai) {
         psm = _psm;
-        gem = _gem;
+        dai = _dai;
     }
 
     function _msgSender() internal pure override returns (address) {
@@ -148,11 +148,11 @@ contract MakerPSMDummy is MakerPSM {
     }
 
     function sellToPool(address recipient, uint256 bps) public {
-        super.sellToMakerPsm(recipient, bps, false, 0, psm, gem);
+        super.sellToMakerPsm(recipient, bps, false, 0, psm, dai);
     }
 
     function buyFromPool(address recipient, uint256 bps) public {
-        super.sellToMakerPsm(recipient, bps, true, 0, psm, gem);
+        super.sellToMakerPsm(recipient, bps, true, 0, psm, dai);
     }
 }
 
@@ -166,7 +166,7 @@ contract MakerPSMUnitTest is Utils, Test {
     address USDC = _etchNamedRejectionDummy("USDC", 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48);
     address USDS = _etchNamedRejectionDummy("USDS", 0xdC035D45d973E3EC169d2276DDab16f1e407384F);
     address PSM = LITE_PSM;
-    address GEM = DAI;
+    address PSM_DAI = DAI;
 
     function setUp() public virtual {
         _mockExpectCall(
@@ -182,7 +182,7 @@ contract MakerPSMUnitTest is Utils, Test {
             address(USDC), abi.encodeWithSelector(IERC20.approve.selector, SKY_PSM, type(uint256).max), abi.encode(true)
         );
         _mockExpectCall(address(USDC), abi.encodeWithSelector(IERC20.decimals.selector), abi.encode(6));
-        psm = new MakerPSMDummy(IPSM(PSM), IERC20(GEM));
+        psm = new MakerPSMDummy(IPSM(PSM), IERC20(PSM_DAI));
     }
 
     function testMakerPSMBuy() public {
@@ -190,9 +190,9 @@ contract MakerPSMUnitTest is Utils, Test {
         uint256 amount = 99999;
 
         _mockExpectCall(
-            GEM, abi.encodeWithSelector(IERC20.balanceOf.selector, address(psm)), abi.encode(amount * 1 ether / 1e6)
+            PSM_DAI, abi.encodeWithSelector(IERC20.balanceOf.selector, address(psm)), abi.encode(amount * 1 ether / 1e6)
         );
-        //_mockExpectCall(GEM, abi.encodeWithSelector(IERC20.allowance.selector, address(psm), PSM), abi.encode(amount));
+        //_mockExpectCall(PSM_DAI, abi.encodeWithSelector(IERC20.allowance.selector, address(psm), PSM), abi.encode(amount));
         //_mockExpectCall(USDC, abi.encodeWithSelector(IERC20.decimals.selector), abi.encode(6));
         _mockExpectCall(PSM, abi.encodeWithSelector(IPSM.tout.selector), abi.encode(100));
         _mockExpectCall(PSM, abi.encodeWithSelector(IPSM.buyGem.selector, RECIPIENT, 99998), abi.encode(amount));
@@ -216,7 +216,7 @@ contract MakerPSMUnitTest is Utils, Test {
 contract MakerSkyPSMUnitTest is MakerPSMUnitTest {
     function setUp() public override {
         PSM = SKY_PSM;
-        GEM = USDS;
+        PSM_DAI = USDS;
         super.setUp();
     }
 }